the emperor wanted to thank the inventor
the inventor asked for one grain of rice for the first square, 2 for the second,4 for the third, and so on m doubling for each of the 64 squares.
Write a program to calculate how many squares are required to give the inventor at least 1000 grains of rice, at least 1,000,000 grains and at least 1,000,000,000 grains. Youll need a loop,ofcourse, and probably an INT to keep track of which square you are at, an INT to keep the number of grains on the current square you are at an INT to keep track of the grains on all previous squares, We suggest that you write out the value of all your variables for each iteration of the loop so that you can see whats going on.